requirements.txt
pandas
sqlalchemy
snowflake-connector-python
python-dotenv
pytest
